+++ packages/implementations/react/src/composite/Accordion/src/components/Body/body.types.ts
@@ -1,27 +1,16 @@
 import { HTMLAttributes, ReactNode } from "react";
-import { accordion } from "@rarui/styles";
+import { AccordionVariants } from "@rarui/styles";
 
-export interface BodyProperties {
+export interface AccordionBodyTyping {
   /**
    * The content of the accordion body.
    * @TJS-type React.ReactNode
    */
   children: ReactNode;
-  /**
-   * The borderBottom property defines a lower border of the accordion body.
-   * @default none
-   */
-  borderBottom?: keyof typeof accordion.properties.borderBottom;
-  /**
-   * The borderTop property defines a top border of the accordion body.
-   * @default none
-   */
-  borderTop?: keyof typeof accordion.properties.borderTop;
-  /**
-   * Padding properties are used to generate space around the content area of an Accordion.Body..
-   * @default base
-   */
-  padding?: keyof typeof accordion.properties.padding;
 }
 
-export type BodyProps = BodyProperties & HTMLAttributes<HTMLElement>;
+export type AccordionBodyProperties = AccordionBodyTyping &
+  Pick<AccordionVariants, "padding">;
+
+export type AccordionBodyProps = AccordionBodyProperties &
+  Omit<HTMLAttributes<HTMLElement>, "color">;