+++ packages/implementations/react/src/composite/Accordion/src/components/Header/index.ts
@@ -1,4 +1,4 @@
-import { Header } from "./Header";
+import { AccordionHeader } from "./Header";
 
-export { Header } from "./Header";
-export default Header;
+export { AccordionHeader } from "./Header";
+export default AccordionHeader;