+++ packages/implementations/react/src/composite/Accordion/src/components/Item/Item.tsx
@@ -1,13 +1,19 @@
 import React, { useMemo } from "react";
 
-import { ManagerItem } from "../../contexts";
-import { ItemProps } from "./item.types";
+import { AccordionManagerItem } from "../../contexts";
+import { AccordionItemProps } from "./item.types";
 
-const Item: React.FC<ItemProps> = ({ children, index, testId }) => {
+const AccordionItem: React.FC<AccordionItemProps> = ({
+  children,
+  index,
+  testId,
+}) => {
   const context = useMemo(() => ({ index, testId }), [index, testId]);
   return (
-    <ManagerItem.Provider value={context}>{children}</ManagerItem.Provider>
+    <AccordionManagerItem.Provider value={context}>
+      {children}
+    </AccordionManagerItem.Provider>
   );
 };
 
-export { Item };
+export { AccordionItem };