+++ packages/implementations/react/src/composite/Accordion/src/contexts/Manager/Manager.tsx
@@ -1,4 +1,6 @@
 import { createContext } from "react";
-import { ManagerProps } from "./manager.types";
+import { AccordionManagerProps } from "./manager.types";
 
-export const Manager = createContext<ManagerProps>({} as ManagerProps);
+export const AccordionManager = createContext<AccordionManagerProps>(
+  {} as AccordionManagerProps,
+);