+++ packages/implementations/react/src/composite/Accordion/src/contexts/ManagerItem/ManagerItem.tsx
@@ -1,6 +1,6 @@
 import { createContext } from "react";
-import { ManagerItemProps } from "./managerItem.types";
+import { AccordionManagerItemProps } from "./managerItem.types";
 
-export const ManagerItem = createContext<ManagerItemProps>(
-  {} as ManagerItemProps,
+export const AccordionManagerItem = createContext<AccordionManagerItemProps>(
+  {} as AccordionManagerItemProps,
 );