+++ packages/implementations/react/src/composite/Accordion/src/contexts/ManagerItem/managerItem.types.ts
@@ -1,4 +1,4 @@
-export interface ManagerItemProps {
+export interface AccordionManagerItemProps {
   index: string;
   testId?: string;
 }