+++ packages/implementations/react/src/composite/Accordion/CHANGELOG.md
@@ -2,12 +2,18 @@
 
 The Accordion component allows the user to show and hide content sections on a page.
 
-## 2024-06-06 `1.0.0`
+## 2024-06-10 `1.0.0`
 
 ### 🎉 New features
 
-- Added `direction` property to the component. ([#51](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/51) by [@junior](https://git.rarolabs.com.br/junior))
-- Added stories to component. ([#51](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/51) by [@junior](https://git.rarolabs.com.br/junior))
+- Added `children` and `selectedDefault` properties to the Component. ([#52](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/52) by [@junior](https://git.rarolabs.com.br/junior))
+- Added stories on Component. ([#52](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/52) by [@junior](https://git.rarolabs.com.br/junior))
+- Created new `Accordion.Header` subcomponent. ([#52](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/52) by [@junior](https://git.rarolabs.com.br/junior))
+- Added `children`, `title`, `subtitle`, `icon`, `noIconToggle` and `size` properties to the Component `Accordion.Header`. ([#52](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/52) by [@junior](https://git.rarolabs.com.br/junior))
+- Created new `Accordion.Body` subcomponent. ([#52](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/52) by [@junior](https://git.rarolabs.com.br/junior))
+- Added `children` and `padding` properties to the Component `Accordion.Body`. ([#52](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/52) by [@junior](https://git.rarolabs.com.br/junior))
+- Created new `Accordion.Item` subcomponent. ([#52](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/52) by [@junior](https://git.rarolabs.com.br/junior))
+- Added `children`, `index` and `testId` properties to the Component `Accordion.Item`. ([#52](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/52) by [@junior](https://git.rarolabs.com.br/junior))
 
 <!-- #### 🛠 Breaking changes -->